+++ test/dummy_app/config/application.rb
@@ -6,9 +6,7 @@ Bundler.require
 
 module Dummy
   class Application < Rails::Application
-    if Rails::VERSION::STRING.to_f >= 5.1
-      config.load_defaults Rails::VERSION::STRING.to_f
-    end
+    config.load_defaults Rails::VERSION::STRING.to_f
 
     # Settings in config/environments/* take precedence over those specified here.
     # Application configuration should go into files in config/initializers